+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename `component` scaffolding to `template`
 * Change default component type for `component` template as `SFC` 
 * Add `documentation` for `component` template
+* Add `size-limit` to templates
 
 ## 3.1.2
 * Fix active class for links in `admin-panel` template